Moved dispatchable to its own header.
